move all servers to the flow folder
