+++ uniflow/flow/config.py
@@ -1,9 +1,138 @@
-"""Flow config module."""
+"""Model, Extract, Transform and Pipeline config module."""
 
-from dataclasses import dataclass
-from uniflow.op.config import ExtractConfig, TransformConfig
+from dataclasses import dataclass, field
+from typing import Dict, Optional
+from uniflow.op.prompt_schema import GuidedPrompt
 
 
+###########################################################
+#                  All Model Configs                      #
+###########################################################
+
+@dataclass
+class ModelConfig:
+    """Model Config Class."""
+
+    model_name: str = "gpt-3.5-turbo-1106"
+    model_server: str = "OpenAIModelServer"
+
+
+@dataclass
+class OpenAIModelConfig(ModelConfig):
+    """OpenAI Model Config Class."""
+
+    model_server: str = "OpenAIModelServer"
+    num_call: int = 1
+    temperature: float = 0.9
+    response_format: Dict[str, str] = field(default_factory=lambda: {"type": "text"})
+
+
+@dataclass
+class HuggingfaceModelConfig(ModelConfig):
+    """Huggingface Model Config Class."""
+
+    model_name: str = "mistralai/Mistral-7B-Instruct-v0.1"
+    batch_size: int = 1
+    model_server: str = "HuggingfaceModelServer"
+
+
+@dataclass
+class LMQGModelConfig(ModelConfig):
+    """LMQG Model Config Class."""
+
+    model_name: str = "lmqg/t5-base-squad-qg-ae"
+    batch_size: int = 1
+    model_server: str = "LMQGModelServer"
+
+
+@dataclass
+class NougatModelConfig(ModelConfig):
+    """Nougat Model Config Class."""
+
+    model_name: str = "0.1.0-small"
+    batch_size: int = 1
+    model_server: str = "NougatModelServer"
+
+
+###########################################################
+#                   All Extract Configs                   # 
+###########################################################
+
+@dataclass
+class ExtractConfig:
+    """Extract Config Class."""
+
+    flow_name: str
+    num_thread: int = 1
+    model_config: Optional[ModelConfig] = None
+
+
+@dataclass
+class ExtractTxtConfig(ExtractConfig):
+    """Extract Txt Config Class."""
+
+    flow_name: str = "ExtractTxtFlow"
+
+
+@dataclass
+class ExtractPDFConfig(ExtractConfig):
+    """Nougat Config Class."""
+
+    flow_name: str = "ExtractPDFFlow"
+    model_config: ModelConfig = NougatModelConfig()
+    
+
+###########################################################
+#                   All Transform Configs                 #    
+###########################################################
+
+@dataclass
+class TransformConfig:
+    """Transform Base Config Class."""
+
+    flow_name: str
+    model_config: ModelConfig
+    num_thread: int = 1
+    guided_prompt_template: GuidedPrompt = GuidedPrompt()
+
+
+@dataclass
+class TransformOpenAIConfig(TransformConfig):
+    """Transform OpenAI Config Class."""
+
+    flow_name: str = "TransformOpenAIFlow"
+    model_config: ModelConfig = OpenAIModelConfig()
+
+
+@dataclass
+class TransformHuggingFaceConfig(TransformConfig):
+    """Transform Hugging Face Config Class."""
+
+    flow_name: str = "TransformHuggingFaceFlow"
+    model_config: ModelConfig = HuggingfaceModelConfig()
+
+
+@dataclass
+class TransformLMQGConfig(TransformConfig):
+    """Transform LMQG Config Class."""
+
+    flow_name: str = "TransformLMQGFlow"
+    guided_prompt_template: GuidedPrompt = GuidedPrompt(instruction="", examples=[])
+    model_config: ModelConfig = OpenAIModelConfig()
+
+
+@dataclass
+class TransformLinearConfig(TransformConfig):
+    """Transform Linear Config Class."""
+
+    flow_name: str = "TransformLinearFlow"
+    guided_prompt_template: GuidedPrompt = GuidedPrompt(instruction="", examples=[])
+    model_config: ModelConfig = field(default_factory=lambda: {})
+    
+    
+###########################################################
+#                    Pipeline Config                      #
+###########################################################
 @dataclass
 class PipelineConfig:
     """Pipeline Config Class."""
