--- uniflow/flow/extract/load/__init__.py
@@ -1,1 +1,0 @@
-"""Flow __init__ module."""
